[CSPM] move k8s specific code to build tagged files
